Switch to curl as the preferred http client
Without this patch wget is the preferred client.  This is a problem
because wget on a CentOS 6 host returns the following error talking to
Github:

    $ wget https://raw.github.com/puppetlabs/puppetlabs-sshkeys/master/templates/scripts/manage_root_authorized_keys
    ERROR: certificate common name “*.a.ssl.fastly.net” doesn’t match requested host name “raw.github.com”.

Curl does not have this issue and is often installed along side wget.
This patch prefers curl in an effort to reduce the likelihood of the
issue.