(BL-36) Do not clobber authorized_keys if download fails
Without this patch the `manage_root_authorized_keys` script will
overwrite the contents of `~root/.ssh/authorized_keys` in the event of a
download failure.  This is a problem because end users cannot log into
new elastic instances.  This patch addresses the problem by checking the
exit status of the download command and aborting the script before
writing to the authorized_keys file if the download does not complete
with a 0 exit status code.