(maint) Fix stray authorized_keys.tmp file left over
Without this patch the manage_root_authorized_keys script leaves a
temporary file in the current working directory of the process.  This is
a problem because the incorrect temporary file is being moved into place
if the script is not executed with ~root/.ssh/ as the current working
directory.

This patch addresses the problem by explicitly writing the temporary
file to the correct location.