(maint) Manage Soalris 10 keys on boot using SMF
Without this patch Solaris 10 VM's cloned from a template may not have
the most up to date copy of the SSH public keys.  This is a problem
because the copy of the authorized_keys sealed in the template may not
be up to date.

This patch addresses the problem by providing an easy way to install an
SMF script on Solaris 10 which manages the root authorized keys during
the boot process.

The repository should be checked out onto the solaris 10 system and
installed using the `install` script prior to sealing the VM template.